# critical realize for unjitted llama