performance critical `realize` for the unjitted path

this PR
```
ran model in 161.57 ms
sync in 4.14 ms
2
ran model in 162.44 ms
sync in 3.56 ms
0
ran model in 162.07 ms
sync in 3.94 ms
 year
ran model in 172.77 ms
sync in 3.67 ms
 old
ran model in 162.81 ms
sync in 3.87 ms
 male
```

master
```
ran model in 190.39 ms
sync in 6.33 ms
2
ran model in 234.34 ms
sync in 4.78 ms
0
ran model in 214.29 ms
sync in 4.13 ms
 year
ran model in 227.56 ms
sync in 4.87 ms
 old
ran model in 232.43 ms
sync in 5.71 ms
 male
```